mathjs library implemented
javascript is not so good at math calculation. javascript calculates: 0.1 + 0.2 = 0.30000000000000004
and many more calculation issues. I've tried my best to solve calculation issues using vanilla javascript but I realized that it can't be solved by javascript.
Because it's not javascript's issue, it's a problem of modern computer. Computer converts decimals to binary and there all problem starts.
However I asked google for the solution and I found mathjs library and implemented it in this commit